chore: redo account details
