+++ holy-grail-frontend/src/features/Account/AccountPage.tsx
@@ -1,11 +1,14 @@
 import { AccountDetails, ChangePassword, UpdateEmail } from '@features';
-import { useState } from 'react';
+import { useState, useContext, useEffect } from 'react';
 import { createTheme, ThemeProvider } from '@mui/material/styles';
 import { VerticalNav, VerticalNavProps } from '@components';
 import AccountBoxIcon from '@mui/icons-material/AccountBox';
 import DescriptionIcon from '@mui/icons-material/Description';
 import PasswordIcon from '@mui/icons-material/Password';
 import Email from '@mui/icons-material/Email';
+import { useNavigate } from 'react-router-dom';
+import { AlertProps, AlertToast } from '@components';
+import { AuthContext } from '@providers';
 
 import './account.css';
 
@@ -15,6 +18,22 @@ export const AccountPage = () => {
   const [subtitle, setSubtitle] = useState('Change and update your account details here!');
 
   const [renderMenuType, setRenderMenuType] = useState<JSX.Element>(<AccountDetails />);
+
+  const { user, isLoading } = useContext(AuthContext);
+  const navigate = useNavigate();
+
+  useEffect(() => {
+    if (!isLoading) {
+      if (!user) {
+        const alertContentRedirect: AlertProps = {
+          title: 'Please login.',
+          description: 'You need to be logged in to edit your account details.',
+          severity: 'error',
+        };
+        navigate('/login', { state: { alertContent: alertContentRedirect } });
+      }
+    }
+  }, [isLoading, user]);
 
   const navProps: VerticalNavProps[] = [
     {
