+++ holy-grail-frontend/src/features/Account/ChangePassword.tsx
@@ -63,7 +63,7 @@ export const ChangePassword = () => {
           style={{
             display: 'grid',
             gridTemplateColumns: '0.2fr auto',
-            gap: '2vh',
+            gridRowGap: '3vh',
             alignItems: 'center',
             marginTop: '2vh',
           }}
